requirements.txt
together
pandas
numpy
matplotlib
seaborn
openpyxl
python-docx
PyPDF2
pillow
streamlit
python-dotenv